+++ desktop/cmp/input/Select.js
@@ -8,7 +8,7 @@ import debouncePromise from 'debounce-promise';
 import {castArray, isEmpty, isNil, isPlainObject, keyBy, merge, isEqual} from 'lodash';
 import PT from 'prop-types';
 import React from 'react';
-import {createFilter, components} from 'react-select';
+import {components} from 'react-select';
 
 import {HoistInput} from '@xh/hoist/cmp/input';
 import {box, div, hbox, span, fragment} from '@xh/hoist/cmp/layout';
@@ -145,11 +145,24 @@ export class Select extends HoistInput {
          */
         queryBuffer: PT.number,
 
+
+        /**
+         * Function to be used for filtering of values for a given query string  input.
+         * Used for filtering of options provided by `options` prop when `enableFilter` is true.
+         * Not to be confused with `queryFn` prop, used in asynchronous mode.
+         *
+         * Provided function should take an option and a query value and return a boolean.
+         * Defaults to a case-insensitive match on word starts.
+         */
+        filterOption: PT.func,
+
+
         /**
          * Async function to return a list of options for a given query string input.
          * Replaces the `options` prop - use one or the other.
          */
         queryFn: PT.func,
+
 
         /**
          * Escape-hatch props passed directly to react-select. Use with care - not all props
@@ -191,18 +204,10 @@ export class Select extends HoistInput {
     // Managed value for underlying text input under certain conditions
     // This is a workaround for rs-select issue described in hoist-react #880
     @observable inputValue = null;
+    inputValueChangedSinceSelect = false;
     get manageInputValue() {
         return this.filterMode && !this.multiMode;
     }
-
-    // Custom local option filtering to leverage default filter fn w/change to show all options if
-    // input has not been changed since last select (i.e. user has not typed).
-    // Applied only when manageInputValue if true.
-    _inputChangedSinceSelect = false;
-    _defaultLocalFilterFn = createFilter();
-    _localFilterFn = (opt, inputVal) => {
-        return !this.inputValue || !this._inputChangedSinceSelect || this._defaultLocalFilterFn(opt, inputVal);
-    };
 
     constructor(props, context) {
         super(props, context);
@@ -262,6 +267,7 @@ export class Select extends HoistInput {
                 onBlur: this.onBlur,
                 onChange: this.onSelectChange,
                 onFocus: this.onFocus,
+                filterOption: this.filterOption,
 
                 ref: this.reactSelectRef
             };
@@ -269,7 +275,6 @@ export class Select extends HoistInput {
         if (this.manageInputValue) {
             rsProps.inputValue = this.inputValue || '';
             rsProps.onInputChange = this.onInputChange;
-            rsProps.filterOption = this._localFilterFn;
         }
 
         if (this.asyncMode) {
@@ -329,7 +334,7 @@ export class Select extends HoistInput {
     onSelectChange = (opt) => {
         if (this.manageInputValue) {
             this.inputValue = opt ? opt.label : null;
-            this._inputChangedSinceSelect = false;
+            this.inputValueChangedSinceSelect = false;
         }
         this.noteValueChange(opt);
     };
@@ -342,11 +347,11 @@ export class Select extends HoistInput {
         if (this.manageInputValue) {
             if (action === 'input-change') {
                 this.inputValue = value;
-                this._inputChangedSinceSelect = true;
+                this.inputValueChangedSinceSelect = true;
                 if (!value) this.noteValueChange(null);
             } else if (action === 'input-blur') {
                 this.inputValue = null;
-                this._inputChangedSinceSelect = false;
+                this.inputValueChangedSinceSelect = false;
             }
         }
     };
@@ -390,6 +395,25 @@ export class Select extends HoistInput {
     //-------------------------
     // Options / value handling
     //-------------------------
+    filterOption = (opt, inputVal) => {
+        // 1) show all options if input has not changed since last select (i.e. user has not typed)
+        if (this.manageInputValue && (!this.inputValue || !this.inputValueChangedSinceSelect)) {
+            return true;
+        }
+
+        // 2) Use function provided by app
+        const {filterOption} = this.props;
+        if (filterOption) {
+            return filterOption(opt, inputVal);
+        }
+
+        // 3) ..or use default word start search
+        if (!inputVal) return true;
+        if (!opt.label) return false;
+        const regex = new RegExp(`(^|\\W)${inputVal}`, 'i');
+        return regex.test(opt.label);
+    };
+
 
     // Convert external value into option object(s). Options created if missing - this takes the
     // external value from the model, and we will respect that even if we don't know about it.